requirements.txt
amqp==5.2.0
argh==0.31.2
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
black==24.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.3
django-appconf==1.0.6
django-autoslug==1.9.9
django-celery-email==3.0.0
django-cors-headers==4.3.1
django-countries==7.5.1
django-environ==0.11.2
django-filter==24.1
django-phonenumber-field==7.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
flake8==7.0.0
flower==2.0.1
humanize==4.9.0
idna==3.6
isort==5.13.2
kombu==5.3.5
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
phonenumbers==8.13.31
pillow==10.2.0
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
psycopg2==2.9.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.4.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tomli==2.0.1
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
watchdog==4.0.0
wcwidth==0.2.13
uwsgi==2.0.20
django-jazzmin==2.6.0
faker
